Draft: mark function to profile from the CLI
This is a draft that allows to profile some function without having to
modify the source.

It does so by inserting an importhook that will be active only for
module/function pairs that are passed on the command line,
and for the given functions it will insert the `@profile` decorator.

The code is not the best, but I'm mostly opening this to get feedback as
to whether you believe this is an interesting possibility to push
forward, or if it something I should just use for me locally/publish as
a separate package.

You can try it with for example

```
$ python -m kernprof -l  --prof bar:f -v line_profiler/foo.py
```

```
$ cat foo.py
from bar import f, g

def main():
    for i in range(60):
        f()
        g()

main()
```
```
$ cat bar.py
import time

def f():
    5
    time.sleep(0.001)
    7

def g():
    pass
```