Stop requiring a bleeding-edge version of Cython.
Requiring `Cython>=3.0.0a11` opens the door to any later pre-release;
this includes the newly-released 3.0.0b1, which breaks backwards
compatibility or requires the addition of a non-backwards-compatible
compiler flag (`legacy_implicit_noexcept`).  Fixing this would require
code changes and pinning the minimum Cython version to 3.0.0b1.

There is no reason to require a bleeding-edge version of Cython.
Doing so opens the door to installs suddenly failing due to breaking
changes in dependencies, such as this one.

Python 3.12 will break compatibility with Cython 0.29, however, so we
make an exception for that version of Python.  For Python 3.12, we
leave Cython pinned to 3.0.0a11, which is the latest version which
works with that Python but is still compatible with the current code.

For other versions of Python, revert the version bump to bleeding-edge
Cython added in a9448ec3b3720dec352bf6a90b5bfbecaf8d13d1, and only require
0.29.24 or higher.

Fixes: #204.