Pin to Cython==3.0.0a11.
The `Cython>=3.0.0a11` requirement allows pypi to pull in 3.0.0b1,
which adds a non-backward-compatible flag which we now require to
compile.

Pin to Cython 3.0.0a11 explicitly, to prevent installs from failing
when they pull in the latest pre-release of Cython.  We continue to
pin to 3.0.0a11, rather than using non-pre-relases, because of
performance improvements in the Cython 3.x series.

Fixes: #204.